+++ src/main.js
@@ -1,5 +1,6 @@
 const { app, BrowserWindow, ipcMain } = require('electron');
-import { declareHandlers } from './handlers'; 
+import { declareHandlers } from './handlers';
+const isDev = require('electron-is-dev');
 
 // Handle creating/removing shortcuts on Windows when installing/uninstalling.
 if (require('electron-squirrel-startup')) {
@@ -21,8 +22,11 @@ const createWindow = () => {
   // and load the index.html of the app.
   mainWindow.loadURL(MAIN_WINDOW_WEBPACK_ENTRY);
 
-  // Open the DevTools.  
-  mainWindow.webContents.openDevTools();
+  // Open the DevTools.
+  if (isDev) {
+	console.log('Running in development...');
+    mainWindow.webContents.openDevTools();
+  }
 };
 
 // This method will be called when Electron has finished
@@ -34,9 +38,9 @@ app.on('ready', createWindow);
 // for applications and their menu bar to stay active until the user quits
 // explicitly with Cmd + Q.
 app.on('window-all-closed', () => {
-  if (process.platform !== 'darwin') {
+  // if (process.platform !== 'darwin') {
     app.quit();
-  }
+  // }
 });
 
 app.on('activate', () => {