+++ src/renderer/app.jsx
@@ -4,6 +4,12 @@ import { createRoot } from 'react-dom/client';
 import { refresh } from './fcpx-scanner.js';
 import App from './fcpx-react.jsx';
 
+// Import our custom CSS
+import './App.scss'
+
+// Import all of Bootstrap's JS
+import './index.css';
+
 const container = document.getElementById('app');
 const root = createRoot(container); // createRoot(container!) if you use TypeScript