+++ src/renderer/fcpx-renderer.js
@@ -4,7 +4,7 @@ import { checkForBackupDisk, addUserDirectory, refresh } from './fcpx-scanner'
 
 checkForBackupDisk().then( path => {
     if(path){
-        console.log("Backups will be done at "+ path);
+        console.log("Backups will be done at " + path);
     } else {
         console.log("No backup disk.");
     }
@@ -16,3 +16,4 @@ window.myAPI.homedir().then(h => {
 });
 
 addUserDirectory("/Volumes");
+window.myAPI.setTitle("xxx");